+++ XpringKit/XRP/Model/XRPPayment.swift
@@ -9,10 +9,18 @@ public struct XRPPayment: Equatable {
   public let amount: XRPCurrencyAmount
 
   /// The unique address of the account receiving the payment.
+  @available(*, deprecated,
+    message: "Please use destinationXAddress, which encodes both the destination and destinationTag")
   public let destination: Address
 
   /// (Optional) Arbitrary tag that identifies a hosted recipient to pay, or the reason for the payment.
+  @available(*, deprecated,
+    message: "Please use destinationXAddress, which encodes both the destination and destinationTag")
   public let destinationTag: UInt32?
+
+  /// The address and (optional) destination tag of the account receiving the payment, encoded in X-address format.
+  /// - SeeAlso: https://xrpaddress.info
+  public let destinationXAddress: Address?
 
   /// (Optional) Minimum amount of destination currency this transaction should deliver.
   public let deliverMin: XRPCurrencyAmount?