+++ XpringKit/XRP/Model/XRPTransaction.swift
@@ -11,6 +11,7 @@ public struct XRPTransaction: Equatable {
   public let hash: String
 
   /// The unique address of the account that initiated the transaction.
+  @available(*, deprecated, message: "Please use sourceXAddress, which encodes both the account and sourceTag")
   public let account: Address
 
   /// (Optional) Hash value identifying another transaction. If provided, this transaction is only valid if
@@ -45,7 +46,12 @@ public struct XRPTransaction: Equatable {
   /// (Optional) Arbitrary integer used to identify the reason for this payment or a sender on whose behalf this
   /// transaction is made.
   /// Conventionally, a refund should specify the initial payment's SourceTag as the refund payment's DestinationTag.
+  @available(*, deprecated, message: "Please use sourceXAddress, which encodes both the account and sourceTag")
   public let sourceTag: UInt32?
+
+  /// The unique address and source tag of the sender that initiated the transaction, encoded as an X-address.
+  /// - SeeAlso: "https://xrpaddress.info"
+  public let sourceXAddress: Address?
 
   /// The signature that verifies this transaction as originating from the account it says it is from.
   public let transactionSignature: Data